refactor(core,toolkit): relocate customJwt local vm handler
relocate customJwt local vm handler
