--- packages/core/src/utils/custom-jwt/local-vm-error.ts
@@ -1,17 +1,0 @@
-import { ResponseError } from '@withtyped/client';
-
-// Extend the ResponseError from @withtyped/client, so that we can unify the error handling and display logic for both OSS version and Cloud version.
-export class LocalVmError extends ResponseError {
-  constructor(errorBody: Record<string, unknown>, statusCode: number) {
-    super(
-      new Response(
-        new Blob([JSON.stringify(errorBody)], {
-          type: 'application/json',
-        }),
-        {
-          status: statusCode,
-        }
-      )
-    );
-  }
-}
