+++ packages/core/src/utils/custom-jwt/local-vm.ts
@@ -1,4 +1,27 @@
+import { types } from 'node:util';
 import { runInNewContext } from 'node:vm';
+
+import { ResponseError } from '@withtyped/client';
+
+/**
+ * Extend the ResponseError from @withtyped/client.
+ * This class is used to parse the error from the local VM to the WithTyped client response error.
+ * So we can unify the error handling and display logic for both local VM and Cloud version.
+ */
+export class LocalVmError extends ResponseError {
+  constructor(errorBody: Record<string, unknown>, statusCode: number) {
+    super(
+      new Response(
+        new Blob([JSON.stringify(errorBody)], {
+          type: 'application/json',
+        }),
+        {
+          status: statusCode,
+        }
+      )
+    );
+  }
+}
 
 /**
  * This function is used to execute a named function in a customized code script in a local
@@ -39,3 +62,18 @@ export const runScriptFunctionInLocalVm = async (
 
   return result;
 };
+
+/**
+ * Build the error body for the local VM error.
+ *
+ * @remarks
+ *
+ * Catch the error from vm module, and build the error body.
+ * Use `isNativeError` to check if the error is an instance of `Error`.
+ * If the error comes from `node:vm` module, then it will not be an instance of `Error` but can be captured by `isNativeError`.
+ *
+ */
+export const buildLocalVmErrorBody = (error: unknown) =>
+  types.isNativeError(error)
+    ? { message: error.message, stack: error.stack }
+    : { message: String(error) };
