--- packages/toolkit/core-kit/src/custom-jwt/index.ts
@@ -1,2 +1,0 @@
-export * from './script-execution.js';
-export * from './error-handling.js';
